+++ NEWS.md
@@ -1,5 +1,6 @@
-# yulab.utils 0.0.5.001
+# yulab.utils 0.0.6
 
++ `mat2df()` to convert matrix to a tidy data frame (2022-12-20, Tue)
 + `str_starts()` and `str_ends()` (2022-07-29, Fri)
 
 # yulab.utils 0.0.5